feat(payment): implement payment proof upload functionality and order status update
